refactor:Workspace settings
