+++ src/components/HomeComponent.js
@@ -1,21 +1,9 @@
 class HomeComponent {
 
-    get workspaceSettingMenu() { return $('[data-testid=admin-settings-dropdown-button]'); }
-    get workspaceSettingsDropdown() { return $('[data-testid=popover-settings-button]'); }
     get createMenuButton() { return $('[data-testid="header-create-menu-button"]'); }
     get createBoardButton() { return $('[data-testid="header-create-board-button"]'); }
     get boardsLink() { return $('[data-testid="open-boards-link"'); }
     
-
-    async clickworkspaceSettingMenu() {
-        await this.workspaceSettingMenu.waitForDisplayed();
-        await this.workspaceSettingMenu.click();
-    }
-
-    async clickworkspaceSettingsDropdown() {
-        await this.workspaceSettingsDropdown.waitForDisplayed();
-        await this.workspaceSettingsDropdown.click();
-    }
 
     async clickCreateMenu() {
         await this.createMenuButton.waitForDisplayed();
