+++ src/components/board/WorkspaceSettingsComponent.js
@@ -1,5 +1,5 @@
-class WorkspaceSettingsComponents {
-    
+class WorkspaceSettingsComponent {
+
     get editIcon() { return $('span[data-testid=EditIcon]'); }
     get workspaceName() { return $('#displayName'); }
     get saveButton() { return $('button[type=submit]'); }
@@ -24,6 +24,7 @@ class WorkspaceSettingsComponents {
         await this.workspaceNameLabel.waitForDisplayed();
         return await this.workspaceNameLabel.getText();
     }
+    
 }
 
-module.exports = new WorkspaceSettingsComponents();
+module.exports = new WorkspaceSettingsComponent();
