+++ src/pages/WorkspacePage.js
@@ -1,0 +1,12 @@
+const workspaceSettingsComponent = require('../components/board/WorkspaceSettingsComponent');
+const WorkspaceMenuComponent = require('../components/board/WorkspaceMenuComponent');
+
+class WorkspacePage {
+    
+    constructor() {
+        this.workspaceSettings = workspaceSettingsComponent;
+        this.workspaceMenu = WorkspaceMenuComponent;
+    }
+}
+
+module.exports = new WorkspacePage();
