--- src/pages/WorkspaceSettingsPage.js
@@ -1,10 +1,0 @@
-const workspaceSettingsComponents = require('../components/WorkspaceSettingsComponents');
-
-class WorkspaceSettingsPage {
-    
-    constructor() {
-        this.workspaceSettings = workspaceSettingsComponents;
-    }
-}
-
-module.exports = new WorkspaceSettingsPage();
